requirements.txt
duckduckgo-search>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
playwright>=1.47.0
llama_cpp
pymysql
